Move SearchField into component
